Converted Remaining DYNAcore material
The only section of the project that needs transition now is the lightswitches
