+++ config.yml
@@ -7,7 +7,7 @@ controllers:
     name: officeFC
     ip: 10.0.0.100
     system: Fadecandy
-    room: 'office'
+    room: office
 
   bedroomFC:
     name: bedroomFC
